SLS - beginning of Deflection

+++ src/main/java/pl/pawz/zelbet/SLS/Deformation.java
@@ -10,14 +10,14 @@ public class Deformation {
     private float alphaDs1;
     private float alphaDs2;
 
-    public Deformation(float eCm, float eS, float b, float h, float hF,float hFT, float bEff,  float bEffT,float a1, float a2, double aS1, double aS2, float tZero, char longOrShort, float fCm, float fCk, float rH, char cement) {
+    public Deformation(float eCm, float eS, float b, float h, float hF, float hFT, float bEff, float bEffT, float a1, float a2, double aS1, double aS2, double tZero, char longOrShort, float fCm, float fCk, float rH, char cement) {
         this.fCm = fCm;
         this.fCk = fCk;
         this.rH = rH;
         this.cement = cement;
 
 
-        BasicParameters basic = new BasicParameters(eCm, eS, b, h, hF,hFT, bEff,  bEffT,a1,a2,aS1,aS2, fCk, rH, tZero, cement, longOrShort);
+        BasicParameters basic = new BasicParameters(eCm, eS, b, h, hF, hFT, bEff, bEffT, a1, a2, aS1, aS2, fCk, rH, tZero, cement, longOrShort);
 
     }
